ml_summarizer/requirements.txt
bert-extractive-summarizer
torch